backend/all_in_one/requirements.txt
google-cloud-aiplatform
google-cloud-firestore
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
numpy==1.26.4
opencv-python-headless==4.7.0.72
firebase-admin
